test/kubernetes/e2e/features/aiextension/tests/requirements.txt
openai
mistralai
anthropic
google.generativeai
vertexai
pytest
langchain-openai
langchain-community
langchainhub
langchain_postgres
fastapi[standard]
uvicorn[standard]
presidio_analyzer[transformers]
presidio_anonymizer
transformers
shapely>=2